+++ android/src/main/java/studio/midoridesign/gal/GalPlugin.java
@@ -77,7 +77,8 @@ public class GalPlugin implements FlutterPlugin, MethodCallHandler, ActivityAwar
             case "putImageBytes": {
                 new Thread(() -> {
                     try {
-                        putMediaBytes(call.argument("bytes"), call.argument("album"));
+                        putMediaBytes(call.argument("bytes"), call.argument("album"),
+                                call.argument("name"));
                         new Handler(Looper.getMainLooper()).post(() -> result.success(null));
                     } catch (Exception e) {
                         handleError(e, result);
@@ -125,11 +126,12 @@ public class GalPlugin implements FlutterPlugin, MethodCallHandler, ActivityAwar
         }
     }
 
-    private void putMediaBytes(byte[] bytes, String album) throws IOException, SecurityException {
+    private void putMediaBytes(byte[] bytes, String album, String name)
+            throws IOException, SecurityException {
         ImageFormat imageFormat = Imaging.guessFormat(bytes);
         String extension = "." + imageFormat.getDefaultExtension().toLowerCase();
         try (InputStream in = new ByteArrayInputStream(bytes)) {
-            writeData(in, true, "image", extension, album);
+            writeData(in, true, name, extension, album);
         }
     }
 
@@ -137,13 +139,28 @@ public class GalPlugin implements FlutterPlugin, MethodCallHandler, ActivityAwar
             String album) throws IOException, SecurityException, FileNotFoundException {
         ContentResolver resolver = pluginBinding.getApplicationContext().getContentResolver();
         ContentValues values = createContentValues(isImage, name, extension, album);
+        Uri uri = getUniqueFileUri(resolver, values, isImage, name, extension);
 
-        Uri uri = resolver.insert(isImage ? IMAGE_URI : VIDEO_URI, values);
         try (OutputStream out = resolver.openOutputStream(uri)) {
             byte[] buffer = new byte[8192];
             int bytesRead;
             while ((bytesRead = in.read(buffer)) != -1) {
                 out.write(buffer, 0, bytesRead);
+            }
+        }
+    }
+
+    // This is necessary because FileUtils.java only checks up to 31 times to generate unique names.
+    // See: https://github.com/natsuk4ze/gal/issues/198
+    private Uri getUniqueFileUri(ContentResolver resolver, ContentValues values, boolean isImage,
+            String name, String extension) throws IllegalStateException {
+        for (int suffix = 0;; suffix++) {
+            try {
+                values.put(MediaStore.MediaColumns.DISPLAY_NAME,
+                        name + (suffix > 0 ? suffix : "") + extension);
+                return resolver.insert(isImage ? IMAGE_URI : VIDEO_URI, values);
+            } catch (IllegalStateException e) {
+                if (!e.getMessage().contains("Failed to build unique file")) throw e;
             }
         }
     }