+++ lib/src/gal_method_channel.dart
@@ -34,10 +34,11 @@ final class MethodChannelGal extends GalPlatform {
   }
 
   @override
-  Future<void> putImageBytes(Uint8List bytes, {String? album}) async {
+  Future<void> putImageBytes(Uint8List bytes,
+      {String? album, required String name}) async {
     await requestAccess(toAlbum: album != null);
     await _invokeMethod<void>(
-        'putImageBytes', {'bytes': bytes, 'album': album});
+        'putImageBytes', {'bytes': bytes, 'album': album, 'name': name});
   }
 
   @override